fix sclaecube lib bundle
