+++ packages/scalecube-microservice/package.json
@@ -46,7 +46,7 @@
     "rollup-plugin-replace": "^2.2.0",
     "rollup-plugin-typescript2": "^0.21.1",
     "rollup-plugin-uglify-es": "^0.0.1",
-    "rollup-plugin-visualizer": "^2.2.0",
+    "rollup-plugin-visualizer": "^2.6.0",
     "ts-jest": "^23.10.5",
     "tslint": "^5.11.0",
     "typescript": "^3.2.4"
