+++ packages/scalecube-microservice/src/TransportProviders/ConnectionManager.ts
@@ -1,14 +1,14 @@
 // @ts-ignore
-import { RSocketClientSocket } from 'rsocket-core';
+import { ReactiveSocket } from 'rsocket-types';
 import { CreateConnectionManager } from '../helpers/types';
 
 export const createConnectionManager: CreateConnectionManager = () => {
-  const openConnections: { [key: string]: Promise<RSocketClientSocket> } = {};
+  const openConnections: { [key: string]: Promise<ReactiveSocket> } = {};
 
   return {
     getConnection: (connectionAddress: string) => openConnections[connectionAddress],
     getAllConnections: () => ({ ...openConnections }),
-    setConnection: (connectionAddress: string, value: Promise<RSocketClientSocket>) =>
+    setConnection: (connectionAddress: string, value: Promise<ReactiveSocket>) =>
       (openConnections[connectionAddress] = value),
     removeConnection: (connectionAddress: string) => delete openConnections[connectionAddress],
   };
