+++ packages/transport-browser/package.json
@@ -29,13 +29,15 @@
     "rollup-plugin-visualizer": "^2.2.0",
     "ts-jest": "^23.10.5",
     "tslint": "^5.11.0",
-    "typescript": "^3.2.4"
+    "typescript": "^3.2.4",
+    "rollup-plugin-commonjs": "^10.0.1"
   },
   "dependencies": {
     "@scalecube/api": "^0.0.3-next.0",
     "@scalecube/utils": "^0.0.3-next.1",
     "rsocket-core": "^0.0.16",
-    "rsocket-events-client": "^0.0.16-6",
-    "rsocket-events-server": "^0.0.16-6"
+    "rsocket-events-client": "^0.0.19",
+    "rsocket-events-server": "^0.0.19",
+    "rsocket-types": "^0.0.16"
   }
 }
