+++ packages/transport-browser/rollup.config.js
@@ -2,6 +2,7 @@ import visualizer from 'rollup-plugin-visualizer';
 import typescript from 'rollup-plugin-typescript2';
 import filesize from 'rollup-plugin-filesize';
 import pkg from './package.json';
+import commonjs from 'rollup-plugin-commonjs';
 
 export default {
   input: 'src/index.ts',
@@ -15,14 +16,20 @@ export default {
       format: 'es',
     },
   ],
-  external: [...Object.keys(pkg.dependencies || {}), ...Object.keys(pkg.peerDependencies || {})],
+  external: ['@scalecube/api', '@scalecube/utils', 'rsocket-core', 'rsocket-flowable'],
   plugins: [
+    commonjs({
+      namedExports: {
+        'rsocket-types': ['CONNECTION_STATUS'],
+      },
+    }),
     typescript({
       typescript: require('typescript'),
       clean: true,
     }),
     visualizer({
       filename: 'report.html',
+      title: 'transport-browser',
     }),
     filesize(),
   ],
