+++ packages/utils/package.json
@@ -26,15 +26,13 @@
     "jest": "^24.6.0",
     "rollup": "^1.14.6",
     "rollup-plugin-filesize": "^6.1.1",
-    "rollup-plugin-node-globals": "^1.4.0",
     "rollup-plugin-typescript2": "^0.21.1",
     "rollup-plugin-visualizer": "^2.2.0",
     "ts-jest": "^23.10.5",
     "tslint": "^5.11.0",
     "typedoc": "^0.14.2",
-    "typescript": "^3.2.4"
-  },
-  "dependencies": {
+    "typescript": "^3.2.4",
     "@scalecube/api": "^0.0.3-next.0"
-  }
+  },
+  "dependencies": {}
 }
